third_party/blink/web_tests/svg/dynamic-updates/SVGFEDropShadowElement-svgdom-dx-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'dx' property of the SVGFEDropShadowElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS dropShadowElement.dx.baseVal is 0
PASS dropShadowElement.dx.baseVal is 10
PASS successfullyParsed is true

TEST COMPLETE
